// sm_vue_client/src/openapi/system/models/VoloAbpPermissionManagementPermissionGroupDto.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
import type { VoloAbpPermissionManagementPermissionGrantInfoDto } from './VoloAbpPermissionManagementPermissionGrantInfoDto';
export type VoloAbpPermissionManagementPermissionGroupDto = {
    name?: string | null;
    displayName?: string | null;
    displayNameKey?: string | null;
    displayNameResource?: string | null;
    permissions?: Array<VoloAbpPermissionManagementPermissionGrantInfoDto> | null;
};