// sm_vue_client/src/openapi/system/models/VoloAbpPermissionManagementUpdatePermissionsDto.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
import type { VoloAbpPermissionManagementUpdatePermissionDto } from './VoloAbpPermissionManagementUpdatePermissionDto';
export type VoloAbpPermissionManagementUpdatePermissionsDto = {
    permissions?: Array<VoloAbpPermissionManagementUpdatePermissionDto> | null;
};